+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - enhancements
+    + [#47202](https://issues.afinedevelopment.com/issues/47202): sticky table headers on scroll
 - bugs
 
 ## 1.1.2